Gradual text rendering.
Adds a system to push text to the UI with each tick, gradually advancing
the char count as it goes.

The overall design here seems wrong. Need to get a better handle on how
the ECS works so we can model this as a single entity with 2 components.
Currently it's using a structure where there's one Entity with handles
to two other Entities, each with their own components (which seems wrong).

Additionally, we'll want to be able to control the speed of the text,
which means tracking the delta time between each tick. Currently one
char per tick means different people with different speed CPUs will see
the text appear at different rates. By tracking the wall clock, and
measuring the deltas, we should be able to render the text at a fixed
speed.

Fixup initial impl to need 1 entity instead of 3.

Originally, while I was cobbling a solution together based on bits and
pieces from the pong tutorial, I had 3 distinct **entities** in play:

- One with the UiText component used to render text to the screen.
- One with the billboard data component (providing the text and "play head").
- One "wrapper" entity that held references to the other two entities (yuk).

This didn't really match my understanding of how this was all meant to
work, so this diff set out to refactor this to use 1 entity with various
components.

Once the entity refactor was done, I had to update the system to iterate
over any entity matching the combination of components we care about,
namely:

- `BillboardData`
- `UiText`

This is to say, any renderable billboard. With this setup we can
theoretically display > 1 billboards on screen at a time (which could be
useful for situations like 2 or more npcs talking over each other).