Glyph reveals based on time, not ticks. Fixes #6.
This diff uses a new field on `BillboardData` to track how long it's
been since the last time glyphs were revealed, and adds the delta time
to it per tick.

When the "since" field has exceeded the glyph reveal period, we loop
through the reveal code once per mod of the time since and the period.

The mod-sized loop is meant to cover cases where the time between ticks
somehow is larger than the period for a single glyph reveal, so as to
"catch up."

The original writeup for #6 described setting this mechanic up using a
new component, but bleh - I'm just going to keep things sloppy for now
and refactor later. It might be prudent to split things up, but I'll
worry about that later.