Add a fixed dimension graphic for billboard background.
This is sort of the weakest version of this I could think to do. A
better, less naive, version would cut the corners and edges up so that
the rounded corner box could be scaled without distortion, but I'll get
to that later I suppose.

The graphic was drawn in aseprite, and the project file is included
alongside the exported png.

Note that this graphic is loaded as part of the Ui scene rather than as a
sprite render entity. Including it as a part of the Ui scene means it's
easier to anchor/position it along with the text. It also saves us from
having to reconfigure the world camera. When I attempted to render this
as a sprite I noticed the graphic would not be visible unless I
repositioned the camera to z > 0 (probably because by default the sprite
itself is at 0, and so is the camera).

Fixes #5.