Move BillboarDisplaySystem code into Playback game state.
Dedicating a system to the act of rendering dialogue text is
probably overkill. Many aspects of what is happening to render the
speaker name and passage text is very specific and not generalized to
the point it makes sense as a "component" and "system" pair.

The overall flow of the code could use another pass at this point. It's
feeling messy to cram all this into "update" but it'll work for right
now.

The `paused` field has been removed since the `ActionTracker` will soon be
used for different things depending on the active state.