Add PromptState which waits for an action to fire before popping.
The ActionTracker component and system have been updated so that the
input polling logic is now encapsulated by methods on the component
struct. This allows us to continue to use ActionTracker as components
(unsure if this will prove useful in reality) but also in game state
code (as we are doing now).

If the Component and System go unused for much longer, they can be
removed. For now, I "split the baby" by putting the logic in the more
neutral location of struct methods.