Toggle "next_page" icon visibility when entering `PromptState`.
Seems like the way to toggle a Ui element's visibility is to either add
or remove a `Hidden` or `HiddenPropagate` component to the entity.

The clue that nudged me towards this solution was a [PR] to the amethyst
book, which is not yet linked to from the published docs on the web.

Seems like there are a number of enhancements that could be made RE:
state handling, animation, etc. These will be addressed in follow-up
revs.

[PR]: https://github.com/amethyst/amethyst/pull/2311/files