Add optional path to dialogue file as arg, tweak input handling.
As a convenience, we can now feed a path to a dialogue file to the app
so long as it is under the asset directory.

The input handling tweak was designed to prevent "bleed" between the
prompt and playback states, requiring the player to release the action
before it will register again in the next state.