Tidy up input handling in states.
Removes the need for `depressed` fields on prompt and playback states by
adjusting how input is viewed by the state.

Lesson learned: game state fields should be modeled around their semantic
meaning instead of the data that will drive them. Changing the playback
state to have a `fastforward` field instead of a `depressed` field
(which was used to guard against the premature activation of the speedup
effect when transitioning away from the prompt state... the details of
the input state should not have been visible in this way.

With the change to tracking `fastforward` instead of button states, we
are able to adjust the field based on input in more appropriate ways.
Now we set `fastforward` to true only when the button is
not-pressed-then-becomes-pressed. We set to false when the button
changes from pressed to released. Ignoring the actual "pressed" state in
between fixes the "bleed" problem the old `depressed` field aimed to
solve.