Parse dialogue files as toml.
This diff puts the pest parser on ice. For now the parser module is not
going to be stitched into the module graph.
Instead we're using toml/serde to handle the broad stokes.

We'll bring pest back in when we get to the point of sprinkling control
markers into the passage text for controling the appearance and speed of the
text, as well as cuing audio/animations and so on.