Build buttons when ChoiceState starts, remove on stop.
After messing with the UiButtonBuilder for a while, I wasn't able to
figure out how to take the `UiButton` it produces and add it into the
ECS like we do `UiText` and so on. There doesn't appear to be a trivial
way to create objects in the "widget" category without going through a
prefab loader so I ended up going with plan B.

Plan B was to use plain `UiText` components for each button and add in
other components to manage the input handling. There's an `Interactable`
component that allows the entity to fire `UiEvent`s on the game state.

This diff builds buttons for each of the choices on start and removes them
when the state stops. Event handling will be added in the next rev.