Add conversation branching. (#22)
* Add choices to passage groups. Stub of new ChoiceState.

Also adds a toml file for working on the choices feature.

* Build buttons when ChoiceState starts, remove on stop.

After messing with the UiButtonBuilder for a while, I wasn't able to
figure out how to take the `UiButton` it produces and add it into the
ECS like we do `UiText` and so on. There doesn't appear to be a trivial
way to create objects in the "widget" category without going through a
prefab loader so I ended up going with plan B.

Plan B was to use plain `UiText` components for each button and add in
other components to manage the input handling. There's an `Interactable`
component that allows the entity to fire `UiEvent`s on the game state.

This diff builds buttons for each of the choices on start and removes them
when the state stops. Event handling will be added in the next rev.

* add left mouse click as "confirm" input

* Add Goto Resource and wire up event handler for `ChoiceState`.

The new `Goto` resource is now used to stash an Option<String>
associated with a given Choice button.
When a button is activated, the value of the `goto` field is pushed into
the resource. The `PlaybackState` then checks for a value in the
resource on resume.