Add cursor position tracking on ChoiceState.
This diff adds action trackers to ChoiceState and uses them to drive a
cursor representing the selected choice. This will ultimately replace
the mouse event code currently in use to trigger state transitions, but
we're not quite there yet.

In addition to adding the trackers, this also refactors aspects of the
button creation such that we keep a separate vec of the y offsets that
correspond to each choice, and therefore each button. I suspect
there will be a refactor in the future to reduce the overall number of
vecs owned by this struct. For now, it is what it is.

Still todo:

- revising the transition code to listen to the confirm action
  (and not the mouse clicks).
- adding a graphic for the cursor position.