Use sprite sheet for icons.
Instead of exporting individual images for each icon, build `UiImage`
components backed by a `SpriteRender`. This is not currently possible to
do when loading `UiImage`s via prefabs, so we must do this in code.

The sprite sheet data and texture are loaded up front in the loading
state. Entities for each icon are pushed into the world as well, retaining the
original names from the RON data, so the downstream game states can
continue to manipulate them in the same way as before.

The prefab loader uses `HiddenPropagate` instead of `Hidden` for some
reason, but instead of continuing this I've switched to plain `Hidden`.
Seems like `HiddenPropagate` would only be needed for complex assemblies
of Ui elements, but `UiImage` is just a texture... so I guess `Hidden` is
fine.