feat: implement Goto mechanic for the choice system
Essentially, this is passing messages from the choice system to the
playback system via a global resource.

There's still a thought that it might be better to handle the Goto
passage setting via a distinct separate game state or system, but for
now I'm handling it all at the head of the playback system.

The difference between this new version and the original amethyst version
is in amethyst the systems had separate methods for doing distinct things
when a game state was paused/resumed. Here there's not really any such
distinction, so having a new state could be a way to represent this.

For now, I leave it. Planning to do some widespread refactor once all
the pieces are in place anyway.