feat: wire up Prompt system to advance through passage groups
The prompt cursor here is prototyped with a red rect, just as was done
for the choice list.
In a future diff we'll tackle actually loading graphics and using the
asset loader, etc.

Additionally, still TODO is the whole playback speed thing (which sort
of dovetails with this prompt system in that it leverages the same input
action).