refactor: split BillboardData into PlayHead and Bookmark
By splitting the large-ish BillboardData component into 2 I'm hoping to
limit or minimize exposure across systems. Certain systems may need
both, but a select few will need one or the other.

Of specific interest is isolating the "bookmark" bit so we can monitor
it for changes to the current passage or passage group, which will
indicate the secs_since_last_reveal timer should be reset.