chore: remove unused stuff
The amethyst integration code for aseprite was not so interesting.
There's a crate out there with the various structs with serde support I
can rely on instead of these, and certainly the amethyst asset loader
would need to be rebuilt for bevy. I'd expect someone already provided
this tbh.

The images and aseprite sources are not a big concern for now since I'm
more or less moving on to other projects. I don't feel the need to load
the sprites as I did in the original amethyst impl.