Migrate from Amethyst to Bevy (#26)
* wip bevy adoption w/ input handling

* WIP migrate from amethyst to bevy

At this stage, we've got placeholder text redered on the screen as well
as some stubbed out plumbing for game states.

Still unclear as to what this will all look like when finished, but at
this point the game builds and runs.

* flesh out Billboard data, stub out Playback state

* partial portage of playback system

* wip messin' with the UI layout

* bump bevy from 0.8 to 0.9

* wip render the choice list from the bottom up

* choice selection input handling

* snip example comment

* feat: implement Goto mechanic for the choice system

Essentially, this is passing messages from the choice system to the
playback system via a global resource.

There's still a thought that it might be better to handle the Goto
passage setting via a distinct separate game state or system, but for
now I'm handling it all at the head of the playback system.

The difference between this new version and the original amethyst version
is in amethyst the systems had separate methods for doing distinct things
when a game state was paused/resumed. Here there's not really any such
distinction, so having a new state could be a way to represent this.

For now, I leave it. Planning to do some widespread refactor once all
the pieces are in place anyway.

* chore: dupe line

* feat: wire up Prompt system to advance through passage groups

The prompt cursor here is prototyped with a red rect, just as was done
for the choice list.
In a future diff we'll tackle actually loading graphics and using the
asset loader, etc.

Additionally, still TODO is the whole playback speed thing (which sort
of dovetails with this prompt system in that it leverages the same input
action).

* feat: wire up the "fast forward" mechanic during playback

* chore: remove defunct amethyst code

The main objective here is to clear out all the amethyst code which has
been ported (naively). The newly ported code will likely get moved
around into modules as a next step. Seems like having this all in one
crate is fine for now, rather than jumping back into using a separate
one.

The amethyst-aesprite crate still remains where it is; a new binding for
Bevy will need to be written (later).

* refactor: ugly move all talkie-specific code into a plugin

This separates the game-level stuff like setting up the camera and
"configure the timestep" from the talkie stuff.

Things are still a mess, but the main.rs is looking more svelte at least.

* chore: clippy lints

There's a type complexity lint which I've silenced for now. The playback
system where it came up will be refactored and this might not be an
issue after that.

* refactor: smaller plugins

I don't know if this is good or not, but it seems like it could be smart
to group components and systems by topic, with narrow access.

* refactor: move Goto-related stuff into a new plugin

Identified during this refactor is an issue where the "seconds since
last reveal" timer needs to be reset in multiple places, ie anywhere
that transitions back to the playback state. The importance is to reset
it directly before transitioning to the playback state and no earlier.
The defect seen otherwise is a burst of text being revealed all at once
when starting a new passage.

A possible solution is noted in comments in the diff, but to summarize
we could isolate parts of the BillboardData component which signal the
reset is required in their own component. That way, the playback system
can watch this one specific component for changes and reset the timer at
the head of the system.

* refactor: split BillboardData into PlayHead and Bookmark

By splitting the large-ish BillboardData component into 2 I'm hoping to
limit or minimize exposure across systems. Certain systems may need
both, but a select few will need one or the other.

Of specific interest is isolating the "bookmark" bit so we can monitor
it for changes to the current passage or passage group, which will
indicate the secs_since_last_reveal timer should be reset.

* refactor: layout improvements, fixup speaker name rendering

* feat(wasm): adds config for running/serving wasm target

The dynamic bevy feature isn't compatible with the wasm target so this
has been adjusted to be a new feature of `talkie-game` itself. This
means when you want to build or run with `--taget=wasm32-unknown-unknown`,
you should also specify `--no-default-features`.

To use `cargo run` with the wasm target, you should first
`cargo install wasm-server-runner`.

Many useful related tips about bevy and wasm can be found in the
cheatbook: <https://bevy-cheatbook.github.io/platforms/wasm.html>

* feat(assets): add custom asset loader support for Dialogue data files

Updates the loading of dialogue files to happen via the bevy asset
server tech instead of looking at the file system directly.

* chore: cargo update

* feat(wasm): add Makefile to help prepare web builds

In order for the wasm target build to be used, you have to bring
together all the neccessary assets as well as some sort of html doc to
load and run the wasm module.

The Makefile just helps to wrangle these files once the build completes.

* chore: remove unused stuff

The amethyst integration code for aseprite was not so interesting.
There's a crate out there with the various structs with serde support I
can rely on instead of these, and certainly the amethyst asset loader
would need to be rebuilt for bevy. I'd expect someone already provided
this tbh.

The images and aseprite sources are not a big concern for now since I'm
more or less moving on to other projects. I don't feel the need to load
the sprites as I did in the original amethyst impl.